Merged several changes proposed by @DanShockley
[merged]Fixed Issue #7
Full modification list:

[merged]-- API key ==> text
Added a comment about the external file.

[reject]-- API root URL ==> text <-- API root URL is derivable from the API key.
Modified to my Root URL. Perhaps should move this to external config file, as well? (not needed - it's not a secret)

[merged]-- get list members ==> JSON
New function to query the members for a list.

[merged]-- member activity feed ==> JSON
New function to retrieve member activity feed for specified member.

[merged]-- add tag (JSON) ==> JSON
New function to add a tag (aka static segment).

[merged]-- tag search ==> JSON
New function to query for the tags of a list.

[merged]-- segment/tag members_to_add ==> JSON
New function to add members to a specified segment/tag (by id) for specified list.

[reject]-- mcpriv - module version ==> version
Changed the version to 2023-03-30, an ISO8601 date, to avoid potential conflict with original author's numbering.

[merged]-- mcpriv - API call ( JSON {endpoint} ; {curl options} ; {data} )
Updated to include HTTP Status Code in the script result JSON. Restore ErrorCaptureState to what it was before script ran.

[merged]-- mcpriv - get listIDs ==> JSON
For a success, store the "result" as a JSONObject, not a JSONString.

[merged]-- mcpriv - get lists ==> JSON
For a success, store the "result" as a JSONObject, not a JSONString.

[merged]-- mcpriv - get list members (JSON) ==> JSON
New function to query the members for a list.

[merged]-- mcpriv - add/update member (JSON) ==> JSON
For a success, store the "result" as a JSONObject, not a JSONString.
As an alternative to subscriber_hash, can specify the MailChimp contact_id.
If the MailChimp contact_id is specified, the email_address can be empty/unspecified, since we might want to update other info by contact_id, without specifying or changing the email address.
That also means we have a good result if either the email_address or the contact_id match.

[merged]-- mcpriv - get member (JSON) ==> JSON
BUG-FIX (Issue #7 on main project): the variable to compare to the "id" returned should be the $subscriber_hash, not the never-defined $member_id.
Use variable(s) to set the info sent to the API call so they can be seen more easily while developing.
If API gets a 404 in the result, that means the MC member was not found, so return a more helpful error message.
When testing for success, ALSO allow the MailChimp "contact_id" to match the "subscriber_hash", since that is another valid way to get a member.

[merged]-- mcpriv - get member activity feed (JSON) ==> JSON
New function to retrieve member activity feed for specified member.

[merged]-- mcpriv - add tag (JSON) ==> JSON
New function to add a tag (aka static segment).

[merged]-- mcpriv - tag search (JSON) ==> JSON
New function to query for the tags of a list.

[merged]-- mcpriv - segment/tag members_to_add (JSON) ==> JSON
New function to add members to a specified segment/tag (by id) for specified list.